requirements.txt
streamlit>=1.40.0
altair>=5.0.0
pandas>=2.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
